requirements.txt
Flask==2.3.2
flask_socketio==5.3.5
pyserial==3.5
